Silence ServerProtocolViolation errors
In the last change, we added more tracing to ServerProtocolViolation WebExceptions in the Crawler. Unfortunately, the ones we are seeing aren't triggering this code because they do not get far enough to get a Response object. I don't expect that we will be able to get useful information from this telemetry, so add this to the set of dependency failures we silently ignore, after adding a trace.